Rename protein_quest.pdbe_3dbeacons.fetch to protein_quest.pdbe_3dbeacons.search
As retrieve module also does fetching and it maps better to cli
